requirements.txt
requests==2.28.1
python-dotenv==0.19.2
rich==13.3.5
typer==0.9.0
pydantic==2.0.3
typing-extensions>=4.6.1
